chore(proto,connector): use int32 for `page_size` param (#19)
Because

- the connector-backend use `int32` as the type of `page_size` now

This commit

- use `int32` for `page_size`

+++ pkg/service/connector.go
@@ -44,7 +44,7 @@ func (s *service) ProbeConnectors(ctx context.Context, cancel context.CancelFunc
 	logger, _ := logger.GetZapLogger(ctx)
 
 	// the number of connector definitions is controllable, using fix size here should be enough
-	pageSize := int64(1000)
+	pageSize := int32(1000)
 	defResp, err := s.connectorPublicClient.ListConnectorDefinitions(ctx, &connectorPB.ListConnectorDefinitionsRequest{PageSize: &pageSize})
 	if err != nil {
 		return err
